+++ packages/cli/src/commands/upload/index.ts
@@ -1,11 +1,9 @@
 import { Command } from '@oclif/core';
 import {
   AddressApi,
-  ChainNameEnum,
   EvmApi,
 } from '@thepowereco/tssdk';
 
-import ux from 'cli-ux';
 import * as Listr from 'listr';
 import { resolve } from 'path';
 import { color } from '@oclif/color';
@@ -17,7 +15,7 @@ import * as abiJson from '../../config/scStorageAbi.json';
 import { storageScAddress } from '../../config/cli.config';
 
 export default class Upload extends Command {
-  static description = 'Upload application files to storage';
+  static description = 'Upload application files to the storage';
 
   static examples = [
     '$ cd app_dir && pow-up',
@@ -32,18 +30,8 @@ export default class Upload extends Command {
 
     let config: CliConfig = await getConfig();
 
-    if (!config) { // TODO: smart prompt
-      const source = await ux.prompt('Please, enter the source path of your project, ex. "./dist")');
-      await ux.confirm(`Source path = "${resolve(source)}". Continue? (yes/no)`);
-
-      const projectId = await ux.prompt('Please, enter your project id (must be unique in list of your projects)');
-      const address = await ux.prompt('Please, enter your account address, ex. "AA030000174483048139"');
-      const wif = await ux.prompt('Please, enter your account private key (wif)', { type: 'hide' });
-
-      config = {
-        source, projectId, address, wif,
-      };
-      await setConfig(config);
+    if (!config) {
+      config = await setConfig();
     }
 
     this.log(color.whiteBright('Current cli config:'));
@@ -55,7 +43,7 @@ export default class Upload extends Command {
     } = config;
 
     const dir = resolve(source);
-    const storageSc = await EvmApi.build(storageScAddress, ChainNameEnum.first, abiJson.abi);
+    const storageSc = await EvmApi.build(storageScAddress, 1, abiJson.abi);
     let taskId = await storageSc.scGet(
       'taskIdByName',
       [AddressApi.textAddressToEvmAddress(address), projectId],