+++ packages/cli/src/types/task.type.ts
@@ -1,0 +1,19 @@
+export class Task {
+  id: number;
+
+  owner: string;
+
+  name: string;
+
+  hash: number;
+
+  size: number;
+
+  taskTime: number;
+
+  expire: number;
+
+  uploader: number;
+
+  status: number;
+}