+++ packages/tssdk/src/typings.ts
@@ -32,6 +32,13 @@ export interface AccountKey {
   wif: string;
 }
 
+export interface RegisteredAccount {
+  chain: number;
+  address: string;
+  seed: string;
+  wif: string;
+}
+
 export interface ChainBootstrapConfig {
   [key: number] : RawNodes;
 }
@@ -50,4 +57,67 @@ export interface ChainConfig {
   callbackCallDelay: number;
   chainRequestTimeout: number;
   maxNodeResponseTime: number;
+}
+
+export interface ChainSettingsCurrentChain {
+  blocktime: number;
+  minsig: number;
+  allowempty: number;
+  patchsigs: number;
+}
+
+export interface ChainSettingsCurrentAllowBlock {
+  block: number;
+  group: number;
+  last: number;
+}
+
+// export interface ChainSettingsCurrentFeeSettingsParams {
+//   feeaddr: string;
+//   notip: number;
+// }
+
+export interface ChainSettingsGasSettings {
+  [key: string] : ChainSettingsGasSettingsValue;
+}
+
+export interface ChainSettingsGasSettingsValue {
+  gasCurrency: string;
+  gas: number;
+  tokens: number;
+}
+
+export interface ChainSettingsCurrencySettingsValue {
+  feeCurrency: string;
+  base: number;
+  baseextra: number;
+  kb: number;
+}
+
+export interface ChainSettingsCurrencySettings {
+  // params: ChainSettingsCurrentFeeSettingsParams; // TODO: fix it
+  [key: string] : ChainSettingsCurrencySettingsValue;
+}
+
+export interface PathKeys {
+  keys: string[];
+  required: number;
+}
+
+export interface ChainSettingsCurrent {
+  chain: ChainSettingsCurrentChain;
+  allocblock: ChainSettingsCurrentAllowBlock;
+  endless: any;
+  fee: ChainSettingsCurrencySettings;
+  gas: ChainSettingsGasSettings;
+  nosk: number;
+  patchkeys: PathKeys;
+  seq: number;
+}
+
+export interface ChainSettings {
+  chains: number[];
+  keys: any[];
+  nodechain: any;
+  current: ChainSettingsCurrent;
 }